spi: fsl-imx-cspi: add explicit compatible strings and required clock properties
The fsl-imx-cspi binding contains language indicating compatible
strings to be used that is not valid for all supported parts
e.g. Should be "fsl,<soc>-cspi" or "fsl,<soc>-ecspi". Fix this
by enumerating the set of valid compatible strings.

The binding is also missing the clocks/clock-names properties
so document these and the two required ipg and per clocks.

Signed-off-by: Matt Porter <mporter@konsulko.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

[ upstream commit: 7f43a87e06e2f19892ff3f278e0181cbf9901b15 ]