Merge branch 'dsa-next'
Florian Fainelli says:

====================
net: dsa: remove restriction on platform_device

This patch series removes the restriction in DSA to operate exclusively with
platform_device Ethernet MAC drivers when using Device Tree. This basically
allows arbitrary Ethernet MAC drivers to be used now in conjunction with
Device Tree.

The reason was that DSA was using a of_find_device_by_node() which limits
the device_node to device pointer search exclusively to platform_device,
in our case, we are interested in doing a "class" research and lookup the
net_device.

Thanks to Chris Packham for testing this on his platform.

Changes in v2:

- fix build for !CONFIG_OF_NET
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: bf0b211256be342e92d3ee5c5a1fb8d42f3928bb ]