Merge branch 'hv_netvsc-next'
K. Y. Srinivasan says:

====================
hv_netvsc: Eliminate memory allocation in the send path

The network protocol used to communicate with the host is the remote ndis (rndis)
protocol. We need to decorate each outgoing packet with a rndis header and
additional rndis state (rndis per-packet state). To manage this state, we
currently allocate memory in the transmit path. Eliminate this allocation by
requesting additional head room in the skb.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 3271e4f18d9a4ec9ac5568f94444dac08f29c5a0 ]