Merge branch 'macb-next'
Nicolas Ferre says:

====================
net/macb: fixes after big driver update

The recent modifications to the macb driver lead to issues with the probe
function code flow. Here are some attempt to fix them.
This time, some more issues are fixed related to the clock as reported by Boris
Brezillon.

The series is written on top of net-next.

Changes in v2:
- address Cyrille comment about exit condition of queue configuration loop
- add fixes for probe sequence related to clocks
- add ethtool register dump
- fix peripheral version test
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: ec9a82fbb10951b1841d8bbbe599f91e8ab2f67f ]